Add `max_arity` option for IND algorithms
When implementing faida, it was decided to temporarily use the `find_nary` option,
which allows user to either mine only unary dependencies or all nary dependencies.
The `max_arity` option allows user to specify the maximum arity of the dependency
(the maximum number of attributes in one of the sides).